requirements.txt
click
skydive-client
graphviz
pandas
ipython
